# Added performance tests