This is the result from running it locally on my desktop.
```
  Performance: 10202 lines JSON
    * Remove token
        22 ms
    √ Remove token (116ms)
    * Select character
        14 ms
    √ Select character (120ms)
    * Select word
        9 ms
    √ Select word (118ms)
    * Select token
        9 ms
    √ Select token (110ms)
    * Select identifier
        6 ms
    √ Select identifier (117ms)
    * Select line
        8 ms
    √ Select line (103ms)
    * Select sentence
        16 ms
    √ Select sentence (115ms)
    * Select paragraph
        5 ms
    √ Select paragraph (112ms)
    * Select document
        13 ms
    √ Select document (114ms)
    * Select nonWhitespaceSequence
        6 ms
    √ Select nonWhitespaceSequence (106ms)
    * Select string
        194 ms
    √ Select string (294ms)
    * Select map
        164 ms
    √ Select map (246ms)
    * Select collectionKey
        142 ms
    √ Select collectionKey (231ms)
    * Select value
        134 ms
    √ Select value (214ms)
    * Select boundedParagraph
        11931 ms
    √ Select boundedParagraph (12011ms)
    * Select boundedNonWhitespaceSequence
        10908 ms
    √ Select boundedNonWhitespaceSequence (10998ms)
    * Select any surrounding pair
        10670 ms
    √ Select any surrounding pair (10768ms)
  17 passing (36s)

```

## Checklist

- [x] I have added [tests](https://www.cursorless.org/docs/contributing/test-case-recorder/)
- [-] I have updated the [docs](https://github.com/cursorless-dev/cursorless/tree/main/docs) and [cheatsheet](https://github.com/cursorless-dev/cursorless/tree/main/cursorless-talon/src/cheatsheet)
- [-] I have not broken the cheatsheet