@AndreasArvidsson 2024-12-18
@pokey Mind having a look at this? Once this is in I have a baseline to compare the collection items scope handler against.