@AndreasArvidsson 2024-12-23
The tests should pass once https://github.com/cursorless-dev/cursorless/pull/2700 is merged